requirements.txt
pytelegrambotapi
requests
pyyaml